Crate quest_hook_proc_macros

Procedural macros for the quest_hook crate

Macros

unsafe_impl_reference_type

Implements the Type trait for a Rust type that is equivalent to a C# reference type

unsafe_impl_value_type

Implements the Type trait for a Rust type that is equivalent to a C# value type

Attribute Macros

hook

Creates an inline hook at a C# method.