quest_hook_proc_macros

Macro unsafe_impl_reference_type

unsafe_impl_reference_type!() { /* proc-macro */ }

Implements the Type trait for a Rust type that is equivalent to a C# reference type

Note that in order to use this macros, the generic_associated_types feature must be enabled.

The Rust type can optionally name the field holding its Il2CppObject header right after the type, separated by a . (e.g. for GameObject.object). When given, that field is used to implement Deref/DerefMut to Il2CppObject, which in turn is what makes the type pick up the blanket RefType/ObjectExt impls (giving it as_object/as_object_mut/new, and autoderef access to Il2CppObject’s own methods like invoke) for free. Leave it off for reference types that aren’t a plain struct with an Il2CppObject field - e.g. ones that provide Deref/DerefMut some other way - and implement RefType (or the traits it’s derived from) by hand instead.

Safety

The type must hold the guarantees required by the Type trait.

Examples

The basic syntax follows the pattern in <libil2cpp path> for <Rust type> [.<object field>] => <C# type>.

#![feature(generic_associated_types)]

use libil2cpp::Il2CppObject;

#[repr(C)]
struct GameObject {
    object: Il2CppObject,
}

unsafe_impl_reference_type!(in libil2cpp for GameObject.object => UnityEngine.GameObject);

It’s also possible to use this macro with generic types. In this scenario, the once_cell feature must also be enabled.

#![feature(generic_associated_types, once_cell)]

use libil2cpp::{Il2CppArray, Il2CppObject, Type};

#[repr(C)]
struct List<T: Type> {
    object: Il2CppObject,
    items: *mut Il2CppArray<T>,
    size: i32,
}

unsafe_impl_reference_type!(in libil2cpp for List<T>.object => System.Collections.Generic.List<T>);

A class getter can be provided manually.

#![feature(generic_associated_types)]

use libil2cpp::Il2CppObject;

#[repr(C)]
struct MyClass {
    object: Il2CppObject,
}

unsafe_impl_reference_type!(in libil2cpp for MyClass.object => MyNamespace.MyClass {
    my_class_getter()
});

Finally, the namespace and class name can be provided as string literals. In this case, the macro will no change them in any way regardless of generic parameters.

#![feature(generic_associated_types, once_cell)]

use libil2cpp::{Il2CppArray, Il2CppObject, Type};

#[repr(C)]
struct List<T: Type> {
    object: Il2CppObject,
    items: *mut Il2CppArray<T>,
    size: i32,
}

unsafe_impl_reference_type!(in libil2cpp for List<T>.object => "System.Collections.Generic"."List`1"<T>);